+++ src/index.js
@@ -5,7 +5,7 @@ const mutation = require('./mutation');
 const crossover = require('./crossover');
 
 const { generateInitialPopulation } = require('./generateInitialPopulation');
-const { reproduce, reproduceAsync } = require('./reproduce');
+const { reproduce } = require('./reproduce');
 const { stopCondition } = require('./stopConditions');
 const randomSequenceOf = require('./randomSequenceOf');
 const randomPermutationOf = require('./randomPermutationOf');
@@ -21,7 +21,6 @@ module.exports = {
 
   generateInitialPopulation,
   reproduce,
-  reproduceAsync,
   stopCondition,
   logIterationData,
   randomSequenceOf,